mods/predict-rain: significant performance improvements
The latest beta introduces persistent global map icons throughout the
entirety of the rift's lifespan, which can be used to efficiently
simulate lunar hail level.

Currently, only a single rift is supported at this time, as there does
not seem to be a way to determine what global map icons originate from.
Thus, a moonlens brought to a rift could trick the implemented logic
into thinking it is the rift portal.

An array of rifts and periodic turf checks can be used to mitigate this,
at the cost of potential performance. However, this would still be
insignificant compared to the high cost of the previous method.